raspberry/requirements.txt
absl-py==2.1.0
asttokens==3.0.0
astunparse==1.6.3
certifi==2022.12.7
cffi==1.17.1
charset-normalizer==2.1.1
cmake==3.25.0
coloredlogs==15.0.1
comm==0.2.2
contourpy==1.3.1
cycler==0.12.1
debugpy==1.8.9
decorator==5.1.1
executing==2.1.0
filelock==3.13.1
flatbuffers==24.3.25
fonttools==4.55.0
fsspec==2024.2.0
gast==0.6.0
google-pasta==0.2.0
grpcio==1.68.1
h5py==3.12.1
humanfriendly==10.0
idna==3.4
ipykernel==6.29.5
ipython==8.30.0
jedi==0.19.2
Jinja2==3.1.3
jupyter_client==8.6.3
jupyter_core==5.7.2
keras==3.7.0
kiwisolver==1.4.7
libclang==18.1.1
lit==15.0.7
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.3
matplotlib-inline==0.1.7
mdurl==0.1.2
ml-dtypes==0.4.1
mpmath==1.3.0
namex==0.0.8
ncnn==1.0.20240820
nest-asyncio==1.6.0
networkx==3.2.1
numpy==1.26.3
nvidia-cublas-cu11==11.11.3.6
nvidia-cuda-cupti-cu11==11.8.87
nvidia-cuda-nvrtc-cu11==11.8.89
nvidia-cuda-runtime-cu11==11.8.89
nvidia-cudnn-cu11==9.1.0.70
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.3.0.86
nvidia-cusolver-cu11==11.4.1.48
nvidia-cusparse-cu11==11.7.5.86
nvidia-nccl-cu11==2.21.5
nvidia-nvtx-cu11==11.8.86
onnx==1.17.0
onnx-graphsurgeon==0.5.2
onnx2tf==1.22.3
onnxruntime==1.20.1
onnxslim==0.1.42
opencv-python==4.10.0.84
opt_einsum==3.4.0
optree==0.13.1
packaging==24.2
pandas==2.2.3
parso==0.8.4
pexpect==4.9.0
pillow==10.2.0
platformdirs==4.3.6
portalocker==3.0.0
prompt_toolkit==3.0.48
protobuf==3.20.3
psutil==6.1.0
ptyprocess==0.7.0
pure_eval==0.2.3
py-cpuinfo==9.0.0
pybind11==2.13.6
pycparser==2.22
Pygments==2.18.0
pyparsing==3.2.0
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
pyzmq==26.2.0
requests==2.28.1
rich==13.9.4
scipy==1.14.1
seaborn==0.13.2
six==1.16.0
sng4onnx==1.0.4
sounddevice==0.5.1
stack-data==0.6.3
sympy==1.13.1
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorflow==2.18.0
tensorflow-io-gcs-filesystem==0.37.1
termcolor==2.5.0
tf_keras==2.18.0
tflite==2.10.0
tflite-support==0.4.4
torch==2.5.1+cu118
torchaudio==2.5.1+cu118
torchvision==0.20.1+cu118
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
triton==3.1.0
typing_extensions==4.9.0
tzdata==2024.2
ultralytics==8.3.39
ultralytics-thop==2.0.12
urllib3==1.26.13
wcwidth==0.2.13
Werkzeug==3.1.3
wrapt==1.17.0
